add support to jump to the scan_range start scan w/o having to sequentially read through all scans starting from 1
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@1266 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ trunk/comet-ms/CometSearch/CometSearchManager.cpp
@@ -2693,7 +2693,7 @@ bool CometSearchManager::WriteIndexedDatabase(void)
    long lProtCount = 0;
    vector<vector<comet_fileoffset_t>> g_pvProteinsList;
    g_pvProteinsList.clear();
-   for (int i = 0; i < g_pvDBIndex.size(); i++)
+   for (size_t i = 0; i < g_pvDBIndex.size(); i++)
    {
       if (i == 0)
       {
